+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.2.0: Swift Package Manager Compatibility (2022-05-05)
+- added compatibility for Swift Package Manager
+
 ## 1.1.0: Minimum Battery Level (2020-05-28)
 - added the `-v` parameter and the `Awaken::version()` property to print version information
 - added the `-b` parameter to set a minimum battery level that causes the sleep assertion to expire